pyproject.toml
[project]
name = "feedme"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "jinja2>=3.1.5",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.taskipy.tasks]
dev = "uv run --env-file='.env' app/debug.py"
revision = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"

[dependency-groups]
dev = [
    "alembic>=1.14.0",
    "debugpy>=1.8.11",
    "pytest>=8.3.4",
    "taskipy>=1.14.1",
]
